Refactor camera picker and update onChange usage
