+++ MEAL AI/Sources/Views/Camera/CameraPickerView.swift
@@ -1,15 +1,54 @@
 import SwiftUI
+import UIKit
 
 struct CameraPickerView: View {
     var onImagesPicked: ([Data]) -> Void
     var onCancel: () -> Void
 
     var body: some View {
-        CustomCameraView(onImageCaptured: { image in
+        CameraView(onImageCaptured: { image in
             if let data = image.jpegData(compressionQuality: 0.8) {
                 onImagesPicked([data])
             }
         }, onCancel: onCancel)
         .ignoresSafeArea()
+    }
+}
+
+private struct CameraView: UIViewControllerRepresentable {
+    var onImageCaptured: (UIImage) -> Void
+    var onCancel: () -> Void
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(self)
+    }
+
+    func makeUIViewController(context: Context) -> UIImagePickerController {
+        let picker = UIImagePickerController()
+        picker.sourceType = .camera
+        picker.delegate = context.coordinator
+        return picker
+    }
+
+    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
+
+    class Coordinator: NSObject, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+        let parent: CameraView
+
+        init(_ parent: CameraView) {
+            self.parent = parent
+        }
+
+        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
+            if let image = info[.originalImage] as? UIImage {
+                parent.onImageCaptured(image)
+            }
+            picker.dismiss(animated: true)
+        }
+
+        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+            parent.onCancel()
+            picker.dismiss(animated: true)
+        }
     }
 }
