--- MEAL AI/Sources/Views/Camera/CustomCameraView.swift
@@ -1,122 +1,0 @@
-import SwiftUI
-import AVFoundation
-import UIKit
-
-struct CustomCameraView: UIViewControllerRepresentable {
-    var onImageCaptured: (UIImage) -> Void
-    var onCancel: () -> Void
-
-    func makeUIViewController(context: Context) -> CameraViewController {
-        let controller = CameraViewController()
-        controller.onImageCaptured = onImageCaptured
-        controller.onCancel = onCancel
-        return controller
-    }
-
-    func updateUIViewController(_ uiViewController: CameraViewController, context: Context) {}
-}
-
-final class CameraViewController: UIViewController {
-    var onImageCaptured: ((UIImage) -> Void)?
-    var onCancel: (() -> Void)?
-
-    private let session = AVCaptureSession()
-    private let photoOutput = AVCapturePhotoOutput()
-    private let previewLayer = AVCaptureVideoPreviewLayer()
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupSession()
-        setupPreview()
-        setupUI()
-    }
-
-    private func setupSession() {
-        session.beginConfiguration()
-        guard let device = AVCaptureDevice.default(.builtInWideAngleCamera, for: .video, position: .back),
-              let input = try? AVCaptureDeviceInput(device: device) else { return }
-        if session.canAddInput(input) { session.addInput(input) }
-        if session.canAddOutput(photoOutput) { session.addOutput(photoOutput) }
-        session.commitConfiguration()
-        session.startRunning()
-    }
-
-    private func setupPreview() {
-        previewLayer.session = session
-        previewLayer.videoGravity = .resizeAspectFill
-        previewLayer.frame = view.bounds
-        view.layer.addSublayer(previewLayer)
-    }
-
-    private func setupUI() {
-        let captureButton = UIButton(type: .system)
-        captureButton.translatesAutoresizingMaskIntoConstraints = false
-        captureButton.setImage(UIImage(systemName: "circle.fill"), for: .normal)
-        captureButton.tintColor = .white
-        captureButton.addTarget(self, action: #selector(capturePhoto), for: .touchUpInside)
-
-        let cancelButton = UIButton(type: .system)
-        cancelButton.translatesAutoresizingMaskIntoConstraints = false
-        cancelButton.setImage(UIImage(systemName: "xmark"), for: .normal)
-        cancelButton.tintColor = .white
-        cancelButton.addTarget(self, action: #selector(cancel), for: .touchUpInside)
-
-        let galleryButton = UIButton(type: .system)
-        galleryButton.translatesAutoresizingMaskIntoConstraints = false
-        galleryButton.setImage(UIImage(systemName: "photo.on.rectangle"), for: .normal)
-        galleryButton.tintColor = .white
-        galleryButton.addTarget(self, action: #selector(openGallery), for: .touchUpInside)
-
-        view.addSubview(captureButton)
-        view.addSubview(cancelButton)
-        view.addSubview(galleryButton)
-
-        NSLayoutConstraint.activate([
-            captureButton.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            captureButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -20),
-
-            galleryButton.centerYAnchor.constraint(equalTo: captureButton.centerYAnchor),
-            galleryButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-
-            cancelButton.centerYAnchor.constraint(equalTo: captureButton.centerYAnchor),
-            cancelButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20)
-        ])
-    }
-
-    @objc private func capturePhoto() {
-        let settings = AVCapturePhotoSettings()
-        photoOutput.capturePhoto(with: settings, delegate: self)
-    }
-
-    @objc private func cancel() {
-        onCancel?()
-    }
-
-    @objc private func openGallery() {
-        let picker = UIImagePickerController()
-        picker.sourceType = .photoLibrary
-        picker.delegate = self
-        present(picker, animated: true)
-    }
-}
-
-extension CameraViewController: AVCapturePhotoCaptureDelegate {
-    func photoOutput(_ output: AVCapturePhotoOutput, didFinishProcessingPhoto photo: AVCapturePhoto, error: Error?) {
-        if let data = photo.fileDataRepresentation(), let image = UIImage(data: data) {
-            onImageCaptured?(image)
-        }
-    }
-}
-
-extension CameraViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        if let image = info[.originalImage] as? UIImage {
-            onImageCaptured?(image)
-        }
-        picker.dismiss(animated: true)
-    }
-
-    func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
-        picker.dismiss(animated: true)
-    }
-}
